Have to imlement actually setting test to num questions selected and add test 2 features.

+++ app/src/main/java/com/danko/jonas/ui_lab1/Results.java
@@ -54,7 +54,7 @@ public class Results implements Serializable{
         double tS = testSize();
         double mark = (s/tS)*100.0;
 
-        return "You have obtained " + (int)s + "/" + (int)tS + ". " + mark +  "%";
+        return "You have obtained " + (int)s + "/" + (int)tS + ". " + (int)mark +  "%";
     }
 
 }
